ssu_apps/requirements/base.txt
Django==1.5
South==0.7.6
Werkzeug==0.8.3
django-braces==0.2.3
django-crispy-forms==1.2.3
django-dajax==0.9.2
django-dajaxice==0.5.4
django-extensions==1.0.1
